refactor: divide line series brush
