+++ src/charts/areaChart.ts
@@ -8,7 +8,7 @@ import Tooltip from '@src/component/tooltip';
 import Plot from '@src/component/plot';
 import AreaSeries from '@src/component/areaSeries';
 import Axis from '@src/component/axis';
-
+import * as lineSeriesBrushes from '@src/brushes/lineSeries';
 import * as basicBrushes from '@src/brushes/basic';
 import * as axisBrushes from '@src/brushes/axis';
 import * as tooltipBrushes from '@src/brushes/tooltip';
@@ -45,6 +45,6 @@ export default class AreaChart extends Chart<AreaChartOptions> {
     this.componentManager.add(Axis, { name: 'xAxis' });
     this.componentManager.add(Tooltip);
 
-    this.painter.addGroups([basicBrushes, axisBrushes, tooltipBrushes]);
+    this.painter.addGroups([basicBrushes, axisBrushes, tooltipBrushes, lineSeriesBrushes]);
   }
 }
