+++ src/charts/lineChart.ts
@@ -7,7 +7,7 @@ import Tooltip from '@src/component/tooltip';
 import Plot from '@src/component/plot';
 import LineSeries from '@src/component/lineSeries';
 import Axis from '@src/component/axis';
-
+import * as lineSeriesBrushes from '@src/brushes/lineSeries';
 import * as basicBrushes from '@src/brushes/basic';
 import * as axisBrushes from '@src/brushes/axis';
 import * as tooltipBrushes from '@src/brushes/tooltip';
@@ -45,6 +45,6 @@ export default class LineChart extends Chart<LineChartOptions> {
     this.componentManager.add(Tooltip);
     this.componentManager.add(Plot);
 
-    this.painter.addGroups([basicBrushes, axisBrushes, tooltipBrushes]);
+    this.painter.addGroups([basicBrushes, axisBrushes, tooltipBrushes, lineSeriesBrushes]);
   }
 }
